requirements.txt
cffi==1.14.3
ffmpeg-python==0.2.0
future==0.18.2
keyboard==0.13.5
MouseInfo==0.1.3
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==7.2.0
PyAudio==0.2.11
PyAutoGUI==0.9.52
pycparser==2.20
PyGetWindow==0.0.9
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python-xlib==0.28
python3-xlib==0.15
PyTweening==1.0.3
six==1.15.0
